Revert "split env.py and gym.py"
This reverts commit f5540940660f8a2dfaeba830e3ac57cc3ce10421.
