+++ pymjx/Makefile
@@ -20,7 +20,7 @@ requirements:
 	python3 -m pip install -r requirements.txt
 
 pytest: requirements
-	python3 -m pytest --doctest-modules --ignore mjx/gym.py
+	python3 -m pytest --doctest-modules --ignore mjx/env.py
 
 test-cli:
 	echo "From mjlog => mjxproto"
