+++ pymjx/mjx/__init__.py
@@ -2,6 +2,5 @@ from mjx.action import Action
 from mjx.diff import check_equality  # noqa: F401
 from mjx.observation import Observation
 from mjx.state import State
-from mjx.env import MjxEnv
 
-__all__ = ["MjxEnv", "Action", "Observation", "State"]
+__all__ = ["Action", "Observation", "State"]
